requirements.txt

streamlit==1.38.0
google-generativeai==0.8.3
python-dotenv==1.0.1
firebase-admin==6.5.0
requests==2.32.3
matplotlib==3.9.2
pandas==2.2.2
numpy==1.26.4
gTTS==2.5.1
plotly==5.24.1
streamlit-drawable-canvas==0.9.3
streamlit-mic-recorder==0.0.8
Pillow==10.4.0
openpyxl==3.1.5
google-cloud-firestore>=2.15


